Transition to conan 2;
Support for Linux build;
Added dds image format decoder;
Animation for masonry layout when changing geometry;
"C" key to dump currently decoded image;
Some QWK related fixes;
Added settings button in viewer (dummy for now)

+++ CacheViewer.cpp
@@ -29,7 +29,7 @@ QVariant ImageInfoModel::data(const QModelIndex &index, int role) const {
     case LastModifiedRole:
         return info.lastModified;
     case ThumbnailSizeRole:
-        return info.location.thumbnailSize;
+        return (quint64)info.location.thumbnailSize;
     case ImageSizeRole:
         return info.imageSize;
     default:
